requirements.txt
Flask
ollama
Pillow
requests
